Aula 3 - Autenticacao OAuth/JWT c/ github web e mobile(quase rs)

+++ server/src/memories/controller.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
 
-import type { FastifyRequest } from 'fastify'
+import type { FastifyReply, FastifyRequest } from 'fastify'
 
 import { prisma } from '../lib/prisma'
 
@@ -12,53 +12,92 @@ export const validarMemoria = z.object({
   content: z.string(),
   coverUrl: z.string().url(),
   isPublic: z.coerce.boolean().default(false),
-  userId: z.string().uuid(),
 })
 
-export const consultar = async () => {
+export const consultar = async (request: FastifyRequest) => {
   return await prisma.memory.findMany({
+    where: {
+      id: request.user.sub,
+    },
     orderBy: {
       createdAt: 'asc',
     },
   })
 }
 
-export const consultarPorId = async (request: FastifyRequest) => {
+export const consultarPorId = async (
+  request: FastifyRequest,
+  reply: FastifyReply,
+) => {
+  await req.jwtVerify()
   const { id } = validarId.parse(request.params)
-  return await prisma.memory.findUniqueOrThrow({ where: { id } })
+
+  const memory = await prisma.memory.findUniqueOrThrow({
+    where: { id },
+  })
+
+  if (memory.isPublic || memory.userId === request.user.sub) {
+    return memory
+  }
+
+  return reply.status(401).send('Unauthorized')
 }
 
 export const criar = async (request: FastifyRequest) => {
-  const { content, coverUrl, isPublic, userId } = validarMemoria.parse(
-    request.params,
-  )
+  await req.jwtVerify()
+  const { content, coverUrl, isPublic } = validarMemoria.parse(request.params)
 
   return await prisma.memory.create({
     data: {
       content,
       coverUrl,
       isPublic,
-      userId,
+      userId: request.user.sub,
     },
   })
 }
 
-export const atualizar = async (request: FastifyRequest) => {
+export const atualizar = async (
+  request: FastifyRequest,
+  reply: FastifyReply,
+) => {
+  await req.jwtVerify()
   const { id } = validarId.parse(request.params)
   const { content, coverUrl, isPublic } = validarMemoria.parse(request.params)
 
-  return await prisma.memory.update({
-    where: { id },
-    data: {
-      content,
-      coverUrl,
-      isPublic,
+  const memory = await prisma.memory.findFirstOrThrow({
+    where: {
+      id,
     },
   })
+
+  if (memory.isPublic || memory.id === request.user.sub) {
+    return await prisma.memory.update({
+      where: { id },
+      data: {
+        content,
+        coverUrl,
+        isPublic,
+      },
+    })
+  }
+
+  return reply.status(401).send('Unauthorized')
 }
 
-export const excluir = async (request: FastifyRequest) => {
+export const excluir = async (request: FastifyRequest, reply: FastifyReply) => {
+  await req.jwtVerify()
   const { id } = validarId.parse(request.params)
 
-  return await prisma.memory.delete({ where: { id } })
+  const memory = await prisma.memory.findFirstOrThrow({
+    where: {
+      id,
+    },
+  })
+
+  if (memory.id === request.user.sub) {
+    return await prisma.memory.delete({ where: { id } })
+  }
+
+  return reply.status(401).send('Unauthorized')
 }
